refactor: vending machine 컴포넌트 내 context 및 provider 분리하여 context 폴더 내 각 provider 컴포넌트로 생성
