--- vending-machine/db/data.json
@@ -1,7 +1,0 @@
-{
-  "item1": {
-    "name": "콜라",
-    "price": "1000",
-    "stock": "2"
-  }
-}
